working on custom paging code

+++ include/mm/kmalloc.h
@@ -4,7 +4,7 @@
 
 void init_heap(uint32_t heap_begin);
 
-uintptr_t dkmalloc(uint32_t size, uint32_t align);
+uint32_t dkmalloc(uint32_t size, uint32_t align);
 
 // TODO:
 // - kmalloc
